extract confirmDialog code from txRunner and move it to its own file
